LPS-201494 Move ScreenReaderAnnouncer components to layout-js-components-web
