+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/fragments_and_widgets/components/reorder_sets_modal/KeyboardDragAndDropContext.tsx
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+import {ScreenReaderAnnouncer} from '@liferay/layout-js-components-web';
 import {sub} from 'frontend-js-web';
 import React, {
 	Dispatch,
@@ -19,7 +20,6 @@ import React, {
 	useState,
 } from 'react';
 
-import ScreenReaderAnnouncer from '../../../../common/components/ScreenReaderAnnouncer';
 import {
 	DRAG_OVER_POSITIONS,
 	DRAG_OVER_POSITIONS_LABELS,
