+++ modules/apps/layout/layout-js-components-web/package.json
@@ -19,7 +19,8 @@
 		"frontend-js-web": "*",
 		"react": "16.12.0",
 		"react-dnd": "11.1.1",
-		"react-dom": "16.12.0"
+		"react-dom": "16.12.0",
+		"uuid": "9.0.0"
 	},
 	"main": "js/index.js",
 	"name": "@liferay/layout-js-components-web",
