+++ modules/apps/layout/layout-js-components-web/src/main/resources/META-INF/resources/js/components/screen_reader_announcer/ScreenReaderAnnouncer.tsx
