+++ modules/apps/layout/layout-js-components-web/src/main/resources/META-INF/resources/js/index.ts
@@ -14,6 +14,12 @@ export {
 	default as ColorPicker,
 	DEFAULT_TOKEN_LABEL,
 } from './components/color_picker/ColorPicker';
+export {default as ScreenReaderAnnouncer} from './components/screen_reader_announcer/ScreenReaderAnnouncer';
+export {
+	ScreenReaderAnnouncerContext,
+	ScreenReaderAnnouncerContextProvider,
+	ScreenReaderAnnouncerContextType,
+} from './contexts/ScreenReaderContext';
 export {default as SearchForm} from './components/search_form/SearchForm';
 export {default as SearchResultsMessage} from './components/search_results_message/SearchResultsMessage';
 export {
